Support make from this decade (newer than 3.81)
This is a kind of roundabout fix: I can't run "make run-asm-tests" in
rocket-chip, and I believe the reason this is happening is because make-3.82
changed pattern rule matching to by "sort by shortest" as opposed to "sort by
first definition".  I've been unable to verify this, but make-3.82 is trying to
match "mm" when building "mm.riscv.hex" while make-3.81 is trying to match
"mm.riscv" when trying to build "mm.riscv.hex".

Since I can't find a real fix, I've just put every benchmark in a ".d"
directory, which is the standard fix when you can't figure out why there's a
file/directory namespace collision.

I've tried this for me on make-4.1 and make-3.81, but it's so fragile it
deserves further testing.