Fixed bug in scall test when running in machine mode
Since stvec_handler was defined to mtvec_handler *after* using
RVTEST_CODE_BEGIN, the macro still uses stvec_handler. This symbol does
not exist in the end, so that it yields zero. For that reason, mtvec
(stvec in the macro) is not set (see the beqz in line 156 of
env/p/riscv_test.h). This has the consequence that the ecall in the test
jumps to the normal trap_vector, which just jumps to write_tohost to
report an error.

I can only guess what the intention of the test is, but I assume
that the ecall executed in user mode should go to machine mode and
jump to stvec_handler. Therefore, I'm setting mtvec manually in the
test.

The patch works both for rv64mi and rv64si on gem5.