Increase timeouts for multi-spike test.
Between October 13 and October 19, something happened that makes the
multi-spike tests 4 times slower. Rolling back spike, OpenOCD, or
riscv-tests doesn't affect this. Presumably it's due to a kernel or
python change in my Ubuntu system.

I don't have time to look at this right now, so just increase the timeouts. :-(

If I had to guess, there could be a bug in rbb_daisychain.py that wastes
a lot of time.