Make MulticoreRegTest work with real hardware.
It would fail intermittently. We can't guarantee all harts resume
simultaneously. When we let multiple harts run to a breakpoint at the
end of the same loop, one is likely to get there first, and the second
won't make it.

To avoid this problem, run for a short amount of time instead of to a
breakpoint.