[debug] Reserve the trigger in `HwbpManual`
After https://github.com/riscv-collab/riscv-openocd/pull/1111 is merged,
the registers a user wishes to have direct control of should be
reserved.
This is the case in `HwbpManual`.

The test still works with older OpenOCD versions, since no exception
is generated when a command (`riscv reserve_trigger` in this case)
is not found.