[debug] Misc debug fixes (#632)
* debug: test S and U are supported for tdata1

Set MCONTROL_S and/or MCONTROL_U bits only if the hart supports S-mode
and/or U-mode.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: test S is supported in TranslateTest

TranslateTest depends on the S extension being supported.
Without this change, the test fails on targets that do not support S
extension.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: rename honors_tdata1_hmode to honors_tdata1_dmode

This is for the dmode bit in tdata1.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: fixed obsoleted openocd-command names

This suppresses warnings like:
  Warning: Command 'gdb_port' is deprecated, please use 'gdb port' instead.
  Warning: Command 'tcl_port' is deprecated, please use 'tcl port' instead.
  Warning: Command 'telnet_port' is deprecated, please use 'telnet port' instead.

from debug/README.md:

> openocd ..., which should be the latest from https://github.com/riscv/riscv-openocd.git.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: not to throw exception on an expected error

This change makes debug/testlib.py not throw an exception
when an expected error occurs while creating the logs/ directory.

Fixing this is useful when debugging with python
debugger not to break at the exception every time.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: call create() instead of target()

U500Sim.py is the only example of a target that uses VcsSim class.
It does not work without this fix.

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

* debug: VcsSim: make server start string configurable

example: for https://github.com/fjullien/jtag_vpi/blob/master/jtag_common.c#L80

    def create(self):
        return testlib.VcsSim(..., server_stared="^Starting jtag_vpi server:.*, port (\d+)/tcp")

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>

---------

Signed-off-by: Hiroo HAYASHI <24754036+hirooih@users.noreply.github.com>